Package test/providers renamed to resource
