+++ models/workitemtype_test.go
@@ -3,14 +3,15 @@ package models
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/almighty/almighty-core/test/providers"
 	"reflect"
 	"testing"
 	"time"
+
+	"github.com/almighty/almighty-core/resource"
 )
 
 func TestJsonMarshalListType(t *testing.T) {
-	providers.Require(t, providers.UnitTest)
+	resource.Require(t, resource.UnitTest)
 
 	lt := ListType{
 		SimpleType: SimpleType{
@@ -51,7 +52,7 @@ func TestJsonMarshalListType(t *testing.T) {
 }
 
 func TestMarshalEnumType(t *testing.T) {
-	providers.Require(t, providers.UnitTest)
+	resource.Require(t, resource.UnitTest)
 
 	et := EnumType{
 		SimpleType: SimpleType{KindEnum},
@@ -87,7 +88,7 @@ func TestMarshalEnumType(t *testing.T) {
 }
 
 func TestMarshalFieldDef(t *testing.T) {
-	providers.Require(t, providers.UnitTest)
+	resource.Require(t, resource.UnitTest)
 
 	et := EnumType{
 		SimpleType: SimpleType{KindEnum},
@@ -111,7 +112,7 @@ func TestMarshalFieldDef(t *testing.T) {
 }
 
 func TestMarshalRawEnum(t *testing.T) {
-	providers.Require(t, providers.UnitTest)
+	resource.Require(t, resource.UnitTest)
 
 	ret := rawEnumType{
 		BaseType: SimpleType{Kind: KindInteger},
@@ -132,7 +133,7 @@ func TestMarshalRawEnum(t *testing.T) {
 }
 
 func TestMarshalArray(t *testing.T) {
-	providers.Require(t, providers.UnitTest)
+	resource.Require(t, resource.UnitTest)
 
 	original := []interface{}{float64(1), float64(2), float64(3)}
 	bytes, err := json.Marshal(original)
